requirements.txt
argparse
requests
python-dotenv